feat(frontend): update useGasFeeUsd to support multiple types of gas param
The `useGasFeeUsd` hook in the `frontend` package has been updated to handle the `gas` parameter as a `bigint`, `string`, or `number`. Previously, it only accepted a `bigint`. This change allows for more flexibility when passing the gas value to the hook. Additionally, the code now converts the `gas` value to a `BigInt` before setting it in state. This ensures consistent handling of different data types.