fix(frontend): handle transaction failure in useSwapToast
Refactored the useSwapToast hook to properly handle transaction failure scenarios. Added logic to display an error toast if the TX fails.

fix(frontend): do not show failed toast

The toast might not fail if there is an error, so no longer switching the toast to a failed one. This might change in the future if we can find a better way of handling it.